Adding support for running scripts to do discovery, start a resource component, and check its availability.
Adding and updating the it (integration test) module to include tests that verify a script
is called for discovery, for starting a resource component, and for checking the component's
availability. And more importantly, the it module helps make it possible or more easy to
run tests out of one's IDE which can be really useful for debugging.

A new class called Action has been added to the alt-lang-plugin module. An instance of
Action is bound to each script context under the name of 'action'. The action provides
information as to what is executing. For example, if the action type is 'resourcecomponent'
and the action name is 'start', then that this indicates that we are inside of
ResourceComponent.start().