Starting to refactor plugin to add unit tests (which should have been done from the start)
With this commit, I have a unit test in place for AltLangComponent.start(). This is the
initial commit for ScriptExecutorService and ScriptExecutorServiceImpl. ScriptExecutorService
handles the work of executing scripts.  It will later also do some validation and can
also provide hooks for pre- and post-processing of scripts. This is also the initial
commit for ScriptResolverFactory. Currently there is only a single ScriptResolver
implementation, but there will be more as the plugin evolves and the factory will encapsulate
the logic of determining which resolver to use.