[BZ 535289] all server-side only code that ignores resource types. if a inventory report from an agent has a resource of an ignored type, it won't get into inventory (you won't see it in the discovery queue). If a parent is in the queue and you import it, any children won't be imported.
A few issues - if the resource is already in the queue and then you ignore it,
you can never take it out of the queue (it won't get committed, it stays
in the NEW state which means it stays in the queue). You can click the
"IGNORE" button in the queue view, but that's a different ignore.

The agent knows nothing about this. So it will still have NEW resources in
its inventory. They will never get committed (so the components won't
be started) however, they will still show in the inventory view of the
agent (it will use some extra memory for sure - there just won't be
started components because the server will never tell the agent the resources
are committed).