[BZ 801926] - manageRootDir deprecated, supeseded by "compliance". The compliance has now 2 possible values: * full (corresponds to manageRootDir=true, i.e. the default), * filesAndDirectories (corresponds to manageRootDir=false)
The name "full" should convey the fact that the deployment directory is in
full compliance with the contents of the bundle.

The name "filesAndDirectories" should convey the behavior of
manageRootDir=false - i.e. the files and directories in the root dir that
are not present in the bundle are left intact. When there is a directory or
file in the root directory that is both in the deployment directory and
the bundle, the file or directory is made compliant to the contents in the
bundle.

The other two proposed deployment behaviors are "rootDirectoryAndFiles"
and "files", but those are commented out for the moment, because we don't
plan to add support for them in RHQ 4.9.