[BZ 873866] - Minimizing API changes between JON 3.1.2.GA and JON 3.2.0.GA.
The fix for bug 1000006 declared the constructor of
o.r.c.d.configuration.ConfigurationUtility private. It previously had the
default pulic constructor and hence this was a API breaking change.

While the utility class should have had the private constructor in first
place, we cannot break the API, so the public constructor has been
re-introduced with a deprecation notice.